refactor: 추가적인 순환 타입 참조 린트 오류 해결
